api/requirements.txt
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==4.8.0
google-api-python-client==2.85.0